rust: implement Serialize/Deserialize manually for Code.
As it seems, large enums are sometimes problematic, because the compiler
tends not inline the serialize/deserialize function. But implementing
that manually based on a primitive type works and lets us directly
serialize/deserialize Code without manual type conversions.

+++ src/kernel/src/tiles/activities.rs
@@ -411,7 +411,7 @@ impl Activity {
             kif::upcalls::Operation::ACT_WAIT,
             kif::upcalls::ActivityWait {
                 event,
-                error: Code::None as u64,
+                error: Code::None,
                 act_sel,
                 exitcode,
             }
@@ -427,7 +427,7 @@ impl Activity {
             kif::upcalls::Operation::DERIVE_SRV,
             kif::upcalls::DeriveSrv {
                 event,
-                error: Code::from(result) as u64
+                error: Code::from(result)
             }
         );
